requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Faker==4.1.1
flasgger==0.5.4
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-restful-swagger-3==0.1
Flask-SQLAlchemy==2.4.3
Flask-WhooshAlchemy==0.56
flask-whooshalchemy3==0.2.0
flask-whooshee==0.7.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
passlib==1.7.2
pbr==3.1.1
PyJWT==1.4.2
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
text-unidecode==1.3
urllib3==1.25.9
Werkzeug==1.0.1
Whoosh==2.7.4
zipp==3.1.0
